add implementation of market use cases in repository
